requirements.txt
ccxt                  # Conexão com Binance e outras exchanges
pandas                # Manipulação de dados e cálculos matemáticos
python-dotenv         # Para ler o arquivo .env
requests              # Para chamadas diretas de API (Glassnode/Telegram)
python-telegram-bot   # Automação de alertas
matplotlib            # Opcional: para gerar gráficos de performance
